+++ R/utils.R
@@ -29,11 +29,28 @@ make_gen_weights <- function(wtype) {
   fun
 }
 
-#Check if all values are the same
-all_the_same <- function(x) {
-  if (is.numeric(x)) return(abs(max(x) - min(x)) < 1e-9)
+na.rem <- function(x) {
+  #A faster na.omit for vectors
+  x[!is.na(x)]
+}
 
-  length(unique(x)) == 1
+check_if_zero <- function(x) {
+  # this is the default tolerance used in all.equal
+  tolerance <- .Machine$double.eps^0.5
+  abs(x) < tolerance
+}
+
+#Check if all values are the same
+all_the_same <- function(x, na.rm = TRUE) {
+  if (anyNA(x)) {
+    x <- na.rem(x)
+    if (!na.rm) {
+      return(is_null(x))
+    }
+  }
+
+  if (is.numeric(x)) check_if_zero(max(x) - min(x))
+  else all(x == x[1])
 }
 
 #Format percentage for CI labels
@@ -42,17 +59,20 @@ fmt.prc <- function(probs, digits = 3) {
 }
 
 check_index <- function(index, t, several.ok = FALSE) {
-  if (length(index) == 0) return(1L)
+  if (is_null(index)) {
+    return(1L)
+  }
 
-  if (ncol(t) == 1) {
+  if (ncol(t) == 1L) {
     if (!((is.numeric(index) && chk::vld_whole_number(index) && index == 1) ||
           (chk::vld_string(index) && !is.null(colnames(t)) && index == colnames(t)[1]))) {
       chk::wrn("only one statistic is available; ignoring `index`")
     }
+
     return(1L)
   }
 
-  if (!((is.character(index) || is.numeric(index)) && is.null(dim(index)))) {
+  if (!((is.character(index) || is.numeric(index)) && is_null(dim(index)))) {
     if (!several.ok) {
       .err("`index` must be a string or number indicating the name or index of the desired statistic")
     }
@@ -76,6 +96,7 @@ check_index <- function(index, t, several.ok = FALSE) {
         .err("`index` must be a vector of positive integers")
       }
     }
+
     if (any(index > ncol(t))) {
       .err("`index` must be between 1 and ", ncol(t))
     }
@@ -84,70 +105,103 @@ check_index <- function(index, t, several.ok = FALSE) {
     if (is.null(colnames(t))) {
       .err("the estimates don't have names, so `index` must be numeric")
     }
-    if (!all(index %in% colnames(t))) {
-      .err("all entries in `index` must be the names of available statistics to compute.\n  The following are allowed: ", paste(dQuote(colnames(t), FALSE), collapse = ", "))
-    }
+
     index <- match(index, colnames(t))
+
+    if (anyNA(index)) {
+      .err("all entries in `index` must be the names of available statistics to compute.\n  The following are allowed: ", paste(add_quotes(colnames(t), 2L), collapse = ", "))
+    }
   }
 
   index
 }
 
-add_quotes <- function(x, quotes = 2) {
-  if (!isFALSE(quotes)) {
-    if (isTRUE(quotes) || as.integer(quotes) == 2) x <- paste0("\"", x, "\"")
-    else if (as.integer(quotes) == 1) x <- paste0("\'", x, "\'")
-    else stop("'quotes' must be boolean, 1, or 2.")
-  }
-  x
-}
-
-word_list <- function(word.list = NULL, and.or = c("and", "or"), is.are = FALSE, quotes = FALSE) {
+word_list <- function(word.list = NULL, and.or = "and", is.are = FALSE, quotes = FALSE) {
   #When given a vector of strings, creates a string of the form "a and b"
   #or "a, b, and c"
   #If is.are, adds "is" or "are" appropriately
-  L <- length(word.list)
-  word.list <- add_quotes(word.list, quotes)
 
-  if (L == 0) {
+  word.list <- setdiff(word.list, c(NA_character_, ""))
+
+  if (is_null(word.list)) {
     out <- ""
     attr(out, "plural") <- FALSE
+    return(out)
+  }
+
+  word.list <- add_quotes(word.list, quotes)
+
+  L <- length(word.list)
+
+  if (L == 1L) {
+    out <- word.list
+    if (is.are) out <- paste(out, "is")
+    attr(out, "plural") <- FALSE
+    return(out)
+  }
+
+  if (is_null(and.or) || isFALSE(and.or)) {
+    out <- paste(word.list, collapse = ", ")
   }
   else {
-    word.list <- word.list[!word.list %in% c(NA_character_, "")]
-    L <- length(word.list)
-    if (L == 0) {
-      out <- ""
-      attr(out, "plural") <- FALSE
-    }
-    else if (L == 1) {
-      out <- word.list
-      if (is.are) out <- paste(out, "is")
-      attr(out, "plural") <- FALSE
+    and.or <- match_arg(and.or, c("and", "or"))
+
+    if (L == 2L) {
+      out <- sprintf("%s %s %s",
+                     word.list[1L],
+                     and.or,
+                     word.list[2L])
     }
     else {
-      and.or <- match_arg(and.or)
-      if (L == 2) {
-        out <- paste(word.list, collapse = paste0(" ", and.or," "))
-      }
-      else {
-        out <- paste(paste(word.list[seq_len(L-1)], collapse = ", "),
-                     word.list[L], sep = paste0(", ", and.or," "))
-
-      }
-      if (is.are) out <- paste(out, "are")
-      attr(out, "plural") <- TRUE
+      out <- sprintf("%s, %s %s",
+                     paste(word.list[-L], collapse = ", "),
+                     and.or,
+                     word.list[L])
     }
-
   }
-  return(out)
+
+  if (is.are) out <- sprintf("%s are", out)
+
+  attr(out, "plural") <- TRUE
+
+  out
+}
+
+add_quotes <- function(x, quotes = 2L) {
+  if (isFALSE(quotes)) {
+    return(x)
+  }
+
+  if (isTRUE(quotes))
+    quotes <- '"'
+
+  if (chk::vld_string(quotes)) {
+    return(paste0(quotes, x, quotes))
+  }
+
+  if (!chk::vld_count(quotes) || quotes > 2) {
+    stop("`quotes` must be boolean, 1, 2, or a string.")
+  }
+
+  if (quotes == 0L) {
+    return(x)
+  }
+
+  x <- {
+    if (quotes == 1) sprintf("'%s'", x)
+    else sprintf('"%s"', x)
+  }
+
+  x
 }
 
 match_arg <- function(arg, choices, several.ok = FALSE) {
   #Replaces match.arg() but gives cleaner error message and processing
   #of arg.
-  if (missing(arg))
-    .err("No argument was supplied to `match_arg()`.", call. = FALSE)
+  if (missing(arg)) {
+    stop("No argument was supplied to match_arg.")
+  }
+
   arg.name <- deparse1(substitute(arg), width.cutoff = 500L)
 
   if (missing(choices)) {
@@ -156,56 +210,71 @@ match_arg <- function(arg, choices, several.ok = FALSE) {
                     envir = sys.frame(sysP))
   }
 
-  if (is.null(arg))
+  if (is_null(arg)) {
     return(choices[1L])
-
-  if (!is.character(arg))
-    .err(sprintf("The argument to `%s` must be `NULL` or a character vector", arg.name), call. = FALSE)
-
-  if (!several.ok) {
-    if (identical(arg, choices))
-      return(arg[1L])
-    if (length(arg) > 1L)
-      .err(sprintf("The argument to `%s` must be of length 1", arg.name), call. = FALSE)
   }
-  else if (length(arg) == 0)
-    .err(sprintf("The argument to `%s` must be of length >= 1", arg.name), call. = FALSE)
+
+  if (several.ok) {
+    chk::chk_character(arg, x_name = add_quotes(arg.name, "`"))
+  }
+  else {
+    chk::chk_string(arg, x_name = add_quotes(arg.name, "`"))
+    if (identical(arg, choices)) {
+      return(arg[1L])
+    }
+  }
 
   i <- pmatch(arg, choices, nomatch = 0L, duplicates.ok = TRUE)
   if (all(i == 0L))
-    .err(sprintf("The argument to `%s` should be %s %s.",
-                 arg.name, ngettext(length(choices), "", if (several.ok) "at least one of " else "one of "),
-                 word_list(choices, and.or = "or", quotes = 2)),
-         call. = FALSE)
+    .err(sprintf("the argument to `%s` should be %s%s",
+                 arg.name,
+                 ngettext(length(choices), "", if (several.ok) "at least one of " else "one of "),
+                 word_list(choices, and.or = "or", quotes = 2)))
   i <- i[i > 0L]
-  if (!several.ok && length(i) > 1)
-    .err("There is more than one match in `match_arg`")
+
   choices[i]
 }
 
-pkg_caller_call <- function(start = 1) {
-  package.funs <- c(getNamespaceExports(utils::packageName()),
-                    .getNamespaceInfo(asNamespace(utils::packageName()), "S3methods")[,3])
-  k <- start #skip checking pkg_caller_call()
-  e_max <- NULL
-  while(!is.null(e <- rlang::caller_call(k))) {
-    if (!is.null(n <- rlang::call_name(e)) &&
-        n %in% package.funs) e_max <- k
-    k <- k + 1
+is_null <- function(x) {length(x) == 0L}
+is_not_null <- function(x) {!is_null(x)}
+
+pkg_caller_call <- function() {
+  pn <- utils::packageName()
+  package.funs <- c(getNamespaceExports(pn),
+                    .getNamespaceInfo(asNamespace(pn), "S3methods")[, 3])
+
+  for (i in seq_len(sys.nframe())) {
+    e <- sys.call(i)
+
+    if (is_null(n <- rlang::call_name(e))) {
+      next
+    }
+
+    if (n %in% package.funs) {
+      return(e)
+    }
   }
-  rlang::caller_call(e_max)
+
+  NULL
 }
 
-.err <- function(...) {
-  chk::err(..., call = pkg_caller_call(start = 2))
+#chk utilities
+.err <- function(..., n = NULL, tidy = TRUE) {
+  m <- chk::message_chk(..., n = n, tidy = tidy)
+  rlang::abort(paste(strwrap(m), collapse = "\n"),
+               call = pkg_caller_call())
 }
-
-.wrn <- function(..., immediate = TRUE) {
-  if (immediate && isTRUE(all.equal(getOption("warn"), 0))) {
+.wrn <- function(..., n = NULL, tidy = TRUE, immediate = TRUE) {
+  if (immediate && isTRUE(all.equal(0, getOption("warn")))) {
     op <- options(warn = 1)
     on.exit(options(op))
   }
-  chk::wrn(...)
+  m <- chk::message_chk(..., n = n, tidy = tidy)
+  rlang::warn(paste(strwrap(m), collapse = "\n"))
+}
+.msg <- function(..., n = NULL, tidy = TRUE) {
+  m <- chk::message_chk(..., n = n, tidy = tidy)
+  rlang::inform(paste(strwrap(m), collapse = "\n"), tidy = FALSE)
 }
 
 .chk_atomic_vector <- function(x, x_name = NULL) {